pyproject.toml
[tool.poetry]
name = "sc-python"
version = "0.1.0"
description = ""
authors = ["Rob Hamilton <berthamilton@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
boto3 = "^1.35.19"
python-sonarqube-api = "^2.0.5"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-bdd = "^7.2.0"
pytest-timeout = "^2.3.1"
pylint = "^3.2.7"
pytest-cov = "^5.0.0"
moby = "^0.0.3"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = "src"
addopts = "-ra -q -s"
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
